Restructured the project to be more modular. Updated the README.md file to reflect the changes.
