+++ Project1/__init__.py
@@ -1,0 +1,1 @@
+# This file can be empty or contain package initialization code if needed.
