+++ Project1/readme.md
@@ -1,0 +1,27 @@
+# Project1
+
+## Overview
+Project1 is a text-based dungeon game where players can explore rooms, pick up items, and navigate through a dungeon.
+
+## Setup
+1. Clone the repository.
+2. Navigate to the `Project1` directory.
+3. Ensure you have Python installed.
+
+## How to Run
+Run the following command to start the game:
+```bash
+python main.py
+```
+
+## Files
+- `main.py`: Entry point of the application.
+- `game.py`: Contains the `Game` class which handles the main game logic.
+- `player.py`: Contains the `Player` class which represents the player.
+- `room.py`: Contains the `Room` class which represents a room in the dungeon.
+- `item.py`: Contains the `Item` class and the `items` dictionary.
+- `dungeon.py`: Contains the `Dungeon` class which generates the dungeon.
+- `__init__.py`: Marks the directory as a package.
+
+## License
+This project is licensed under the MIT License.
