+++ README.md
@@ -1,14 +1,11 @@
 # Projects
-Projects I use to help me learn to code better
-Hopefully this will be full of stuff!
 
+Projects I use to help me learn to code better. Hopefully, this will be full of stuff!
 
-Project 1: "Dungeon game"
-  This should be a cool dungeon exploring game - I hope to learn how to make random rooms, and potentially create descriptions to match.
-  Player movement and interactions will still be limited to basic text commands for now.
+## Project 1: "Dungeon game"
+This should be a cool dungeon exploring game - I hope to learn how to make random rooms, and potentially create descriptions to match. Player movement and interactions will still be limited to basic text commands for now.
 
-Project 2: "Movielog"
-  This will be a simple application to log movies I've watched. I aim to learn how to work with databases and create a user-friendly interface.
-  Users will be able to add, edit, and delete movie entries, as well as rate and review them.
-  I also plan to try and make it online and work with an API.
-    
+For more details, see the [Project1 README](./Project1/README.md).
+
+## Project 2: "Movielog"
+This will be a simple application to log movies I've watched. I aim to learn how to work with databases and create a user-friendly interface. Users will be able to add, edit, and delete movie entries, as well as rate and review them. I also plan to try and make it online and work with an API.
